pyproject.toml
[project]
name = "antigravity"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "shap>=0.42.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "umap-learn>=0.5.0",
    "streamlit>=1.28.0",
    "nbformat>=5.9.0",
    "python-pptx>=0.6.21",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "statsmodels>=0.14.0",
]

[tool.uv]
# Setup: uv venv --python 3.12 && uv sync
# Note: Python 3.13+ breaks SHAP. Pin to 3.12.

